backend/pyproject.toml
[tool.poetry]
name = "Game Overview"
version = "0.1.0"
description = ""
authors = ["Robert Meissner <robert@robertforpresent.de>"]

[tool.poetry.dependencies]
python = "^3.10"
pyperclip = "^1.9.0"
odfpy = "^1.4.1"
python-dotenv = "^0.21.0"
httpx = "^0.27.2"
matplotlib = "^3.6.2"
pre-commit = "^2.20.0"
streamlit = "^1.37.1"
tqdm = "^4.66.5"
pytest = "^8.3.2"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
pandas = "^1.5.2"
numpy = "1.26.3"
bs4 = "^0.0.2"
pyarrow = "^17.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
